Work around lack of TLS support when building with clang
Work around lack of TLS support when building with clang

When building perf with clang, Android's lack of TLS support
results in an undefined reference to __aeabi_read_tp.

As a workaround, don't use __thread until Bionic adds TLS.

Change-Id: I700b151a38e9a99fd19d616aa840a11c4e466f1a
Signed-off-by: Bernhard Rosenkränzer <Bernhard.Rosenkranzer@linaro.org>

based on Bero's Linaro change, but had to modify for the current paths.